fixed compile warning in rtcore_common.h

+++ include/embree3/rtcore_common.h
@@ -223,7 +223,7 @@ RTC_FORCEINLINE void rtcInitIntersectContext(struct RTCIntersectContext* context
 {
   context->flags = RTC_INTERSECT_CONTEXT_FLAG_INCOHERENT;
   context->filter = NULL;
-  context->instID[0] = -1;
+  context->instID[0] = (unsigned int)-1;
 }
   
 #if defined(__cplusplus)
